+++ CHANGELOG.md
@@ -4,6 +4,29 @@
 
 <!-- next version -->
 
+## v0.1.4
+
+### 🚩 Deprecations 🚩
+
+- `pkg/otlp/metrics`: Deprecate metrics.WithCountSumMetrics translator option. (#23)
+  - metrics.WithCountSumMetrics will now send min and max metrics when available.
+  - Use metrics.WithHistogramAggregations instead.
+  
+
+### 💡 Enhancements 💡
+
+- `pkg/otlp/metrics`: Send min and max for delta histograms and delta exponential histograms when metrics.WithHistogramAggregations is used. (#23)
+
+## v0.1.3
+
+### 🚩 Deprecations 🚩
+
+- `pkg/otlp/attributes`: Deprecate and replace functions which have usePreviewRules parameter in preparation to graduate the exporter.datadog.hostname.preview feature gate to stable. (#21)
+
+### 💡 Enhancements 💡
+
+- `pkg/otlp/metrics`: The OTLP metrics converter now maps OpenTelemetry Go runtime metrics to their Datadog counterparts. (#14)
+
 ## v0.1.2
 
 ### 💡 Enhancements 💡